kpi
kpi-changes in table

+++ BusinessLayer/DTOs/PerformanceKPIDto.cs
@@ -13,7 +13,7 @@ namespace BusinessLayer.DTOs
         public decimal? Weightage { get; set; }
         public string? Target { get; set; }
         public string? Achieved { get; set; }
-        public int? SelfRating { get; set; }
+        public decimal? SelfRating { get; set; }
         public int? ManagerRating { get; set; }
         public string? Remarks { get; set; }
     }
